move bottom nav preview to scaffold
